pyproject.toml
[project]
name = "mybot4qq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "aiosmtplib>=5.1.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "nest-asyncio>=1.6.0",
    "pulsar-client>=3.9.0",
    "py-consul>=1.7.1",
    "python-consul>=1.1.0",
    "pyyaml>=6.0.3",
]
